# packages/openapi/openapi.json
{
  "openapi": "3.0.2",
  "paths": {
...
    "/api/v1/auth/refresh": {
      "post": {
        "description": "Refreshes the user's authentication tokens",
        "summary": "Refresh",
        "tags": ["Authentication"],
        "parameters": [],
        "operationId": "refresh",
        "requestBody": {
          "description": "Body",
          "content": {
            "application/json": {
              "schema": {}
            }
          }
        },
        "responses": {
          "200": {
            "description": "200",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "accessToken": {
                      "type": "string"
                    }
                  },
                  "required": ["accessToken"]
                }
              }
            }
          }
        }
      }
    }
  },
  "info": {
    "version": "1.0.0",
    "title": "Loreline REST API - Documentation",
    "description": "Documentation for the Loreline REST API"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "x-service-token": {
        "type": "apiKey",
        "name": "x-service-token",
        "in": "header"
      }
    }
  }
}